mess of changes to get ./test to pass
