+++ enclone_args/src/proc_args.rs
@@ -369,6 +369,7 @@ pub fn proc_args(mut ctl: &mut EncloneControl, args: &Vec<String>) -> Result<(),
         ("NGRAPH_FILTER", &mut ctl.gen_opt.ngraph_filter),
         ("NGROUP", &mut ctl.clono_group_opt.ngroup),
         ("NIMPROPER", &mut ctl.merge_all_impropers),
+        ("NO_NEWLINE", &mut ctl.gen_opt.no_newline),
         ("NO_UNCAP_SIM", &mut ctl.gen_opt.no_uncap_sim),
         ("NON_CELL_MARK", &mut ctl.clono_filt_opt_def.non_cell_mark),
         ("NOPRINT", &mut ctl.gen_opt.noprint),
