+++ enclone_tools/src/bin/review_main_tests.rs
@@ -6,7 +6,7 @@
 // NOTE: you have to run this from the enclone_main directory.  Otherwise it won't work.
 
 use enclone::misc1::setup_pager;
-use enclone_core::testlist::*;
+use enclone_core::main_testlist::*;
 use enclone_tools::run_test::*;
 use pretty_trace::*;
 use rayon::prelude::*;
