+++ enclone_tools/src/bin/run_last_test.rs
@@ -4,7 +4,7 @@
 //
 // Note that this gives the wrong answer if you haven't run ./build first.
 
-use enclone_core::testlist::*;
+use enclone_core::main_testlist::*;
 use enclone_tools::run_test::*;
 use pretty_trace::*;
