+++ enclone_tools/src/bin/update_all_main_tests.rs
@@ -7,7 +7,7 @@
 //
 // NOTE: you have to run this from the enclone_main directory.  Otherwise it won't work.
 
-use enclone_core::testlist::*;
+use enclone_core::main_testlist::*;
 use enclone_tools::run_test::*;
 use io_utils::*;
 use pretty_trace::*;
